requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests==2.32.3
python-multipart==0.0.9
openai>=1.40.0
aiosmtplib
google-cloud-storage
google-cloud-vision
pandas
openpyxl
APScheduler
PyPDF2
python-docx
meilisearch
extract-msg
pypandoc
zipstream
Pillow
pytesseract
passlib[bcrypt]
python-jose[cryptography]
numpy<2.0
chromadb==1.1.0
stripe==7.0.0
bcrypt>=4.0.0
reportlab
pdfplumber>=0.11.0